+++ CHANGELOG.md
@@ -403,7 +403,7 @@
 * [X] 注册基本的`Converter`
 * [X] 统一命名规则为`xxx`和`xxxOrNull`
 * [X] 更新COPYRIGHT
-* [ ] `<Collection>.deepGetOrElse`和`<Collection>.deepGetOrPut`
+* [X] 实现`<Collection>.deepGetOrElse`
 
 # 长期